Calendar adjustments to match real-life international breaks (England) Adjusting UEFA/AFC club comps again to try and reduce game freezing A-League: added the 4 extra games, and Western Utd no longer re-joins the league Fix: players in UEFA Nations League would stay away forever small fix referencing offset 0xA7 -> WORD instead of DWORD vtable renames for clarity

+++ Egypt/egy_league_cup.cpp
@@ -412,7 +412,7 @@ char egy_league_cup_update(BYTE* _this) {
 	}
 	data->year++;
 	data->current_stage = -1;
-	*((DWORD*)(_this + 0xA7)) = -1;
+	*((WORD*)(_this + 0xA7)) = -1;
 	egy_league_cup_all_teams(_this);
 	DWORD v1 = *(DWORD*)_this;
 	(DWORD*)(*(int(__thiscall**)(BYTE*))(v1 + 0x5C))(_this);
